requirements.txt
pillow==2.6.0 # image editing
